moved design plot to reporting and example fixes
